fix(operations): move hierarchical logic into layout wrapSelector
