+++ src/core/components/operations.jsx
@@ -37,145 +37,88 @@ export default class Operations extends React.Component {
     // Get pertinent options
     let {
       hierarchicalTags,
-      tagSplitterChar,
-    } = getConfigs();
-
-    // Set default tagSplitterChar if necessary
-    tagSplitterChar = tagSplitterChar || /[:|]/;
+    } = getConfigs()
 
     // Get a flat map of tag names to tag info and operations. Note that this will always return a
     // flat list, even if the `hierarchicalTags` option is set to `true`.
     const taggedOps = specSelectors.taggedOperations()
 
+    if(taggedOps.size === 0) {
+      return <h3> No operations defined in spec!</h3>
+    }
+
     // Render either hierarchical or flat depending on config
     if (hierarchicalTags) {
-      // If the `hierarchicalTags` option is set, we want to break down the tags into a deep
-      // hierarchy. We're using a "raw" object for cleanliness here, but later we'll convert that
-      // into an immutable map. Here are the types we're dealing with:
-      //
-      // const operationTagsRaw: TagMap;
-      // type TagMap = { [TagName: string]: TagData };
-      // type TagData = {
-      //   canonicalName: string;
-      //   data: TagInfoAndOperations | null;
-      //   childTags: TagMap;
-      // }
-      // TODO: Explicitly define TagInfoAndOperations
-
-      const operationTagsRaw = {};
-
-      // For each raw tag....
-      taggedOps.map((tagObj, tagName) => {
-        // Split the raw tag name into parts
-        const parts = tagName.split(tagSplitterChar);
-
-        // Set a pointer for use in traversing the hierarchy
-        let current = operationTagsRaw;
-
-        // Iterate through the parts defined by this tag
-        for (let i = 0; i < parts.length; i++) {
-          const part = parts[i];
-
-          // If there's no object defined for the current part, define one with just childTags as an
-          // empty set
-          if (current[part] === undefined) {
-            // Compose canonical name from parts up to this point
-            const canonicalName = parts.reduce(
-              (name, p, j) => ((j > i) ? name : name.concat([p])),
-              []
-            ).join("|");
-            current[part] = {
-              canonicalName,
-              data: null,
-              childTags: {}
-            }
-          }
-
-          // If this is the last part, set data on this object
-          if (i === parts.length - 1) {
-            current[part].data = tagObj;
-          }
-
-          // Move to the next level of the hierarchy before looping around
-          current = current[part].childTags;
-        }
-      });
-
-      // Convert to immutable map
-      const operationTags = Im.fromJS(operationTagsRaw);
       const HierarchicalOperationTag = getComponent("HierarchicalOperationTag")
-      return operationTags.size === 0
-        ? <h3> No operations defined in spec!</h3>
-        :
-        <HierarchicalOperationTag
-          specSelectors={specSelectors}
-          oas3Selectors={oas3Selectors}
-          layoutSelectors={layoutSelectors}
-          layoutActions={layoutActions}
-          getConfigs={getConfigs}
-          getComponent={getComponent}
-          childTags={operationTags}
-          isRoot={true}
-        />
-    } else {
-      const OperationContainer = getComponent("OperationContainer", true)
-      const OperationTag = getComponent("OperationTag")
-      return (
-          <div>
-            {
-              taggedOps.map( (tagObj, tag) => {
-                const operations = tagObj.get("operations")
-                return (
-                  <OperationTag
-                    key={"operation-" + tag}
-                    tagObj={tagObj}
-                    tag={tag}
-                    oas3Selectors={oas3Selectors}
-                    layoutSelectors={layoutSelectors}
-                    layoutActions={layoutActions}
-                    getConfigs={getConfigs}
-                    getComponent={getComponent}
-                    specUrl={specSelectors.url()}>
-                    {
-                      operations.map( op => {
-                        const path = op.get("path")
-                        const method = op.get("method")
-                        const specPath = Im.List(["paths", path, method])
-
-
-                        // FIXME: (someday) this logic should probably be in a selector,
-                        // but doing so would require further opening up
-                        // selectors to the plugin system, to allow for dynamic
-                        // overriding of low-level selectors that other selectors
-                        // rely on. --KS, 12/17
-                        const validMethods = specSelectors.isOAS3() ?
-                              OAS3_OPERATION_METHODS : SWAGGER2_OPERATION_METHODS
-
-                        if(validMethods.indexOf(method) === -1) {
-                          return null
-                        }
-
-                        return <OperationContainer
-                                   key={`${path}-${method}`}
-                                   specPath={specPath}
-                                   op={op}
-                                   path={path}
-                                   method={method}
-                                   tag={tag}
-                                   />
-                      }).toArray()
-                    }
-
-
-                  </OperationTag>
-                )
-              }).toArray()
-            }
-
-            { taggedOps.size < 1 ? <h3> No operations defined in spec! </h3> : null }
-          </div>
-      )
+      return <HierarchicalOperationTag
+        specSelectors={specSelectors}
+        oas3Selectors={oas3Selectors}
+        layoutSelectors={layoutSelectors}
+        layoutActions={layoutActions}
+        getConfigs={getConfigs}
+        getComponent={getComponent}
+        childTags={taggedOps}
+        isRoot={true}
+      />
     }
+
+    const OperationContainer = getComponent("OperationContainer", true)
+    const OperationTag = getComponent("OperationTag")
+    return (
+        <div>
+          {
+            taggedOps.map( (tagObj, tag) => {
+              const operations = tagObj.get("operations")
+              return (
+                <OperationTag
+                  key={"operation-" + tag}
+                  tagObj={tagObj}
+                  tag={tag}
+                  oas3Selectors={oas3Selectors}
+                  layoutSelectors={layoutSelectors}
+                  layoutActions={layoutActions}
+                  getConfigs={getConfigs}
+                  getComponent={getComponent}
+                  specUrl={specSelectors.url()}>
+                  {
+                    operations.map( op => {
+                      const path = op.get("path")
+                      const method = op.get("method")
+                      const specPath = Im.List(["paths", path, method])
+
+
+                      // FIXME: (someday) this logic should probably be in a selector,
+                      // but doing so would require further opening up
+                      // selectors to the plugin system, to allow for dynamic
+                      // overriding of low-level selectors that other selectors
+                      // rely on. --KS, 12/17
+                      const validMethods = specSelectors.isOAS3() ?
+                            OAS3_OPERATION_METHODS : SWAGGER2_OPERATION_METHODS
+
+                      if(validMethods.indexOf(method) === -1) {
+                        return null
+                      }
+
+                      return <OperationContainer
+                                 key={`${path}-${method}`}
+                                 specPath={specPath}
+                                 op={op}
+                                 path={path}
+                                 method={method}
+                                 tag={tag}
+                                 />
+                    }).toArray()
+                  }
+
+
+                </OperationTag>
+              )
+            }).toArray()
+          }
+
+          { taggedOps.size < 1 ? <h3> No operations defined in spec! </h3> : null }
+        </div>
+    )
   }
 }
